double => float changes (breaks file format)

+++ em-tools/em-edit/bezier.c
@@ -40,7 +40,7 @@
 #include "bezier.h"
 #include "worker.h"
 
-void BRZ(struct bezier_t *bezier, double t, double *x, double *y)
+void BRZ(struct bezier_t *bezier, float t, float *x, float *y)
 {
 	double B0 = (1 - t) * (1 - t) * (1 - t);
 	double B1 = 3 * t * (1 - t) * (1 - t);
@@ -52,7 +52,7 @@ void BRZ(struct bezier_t *bezier, double t, double *x, double *y)
 }
 
 
-void deltaBRZ(struct bezier_t *bezier, double t, double *x, double *y)
+void deltaBRZ(struct bezier_t *bezier, float t, float *x, float *y)
 {
 	double B0 = -3 + 6 * t - 3 * t * t;
 	double B1 = 3 - 12 * t + 9 * t * t;
@@ -133,7 +133,7 @@ int bezier_t_divide(struct t_t *t)
 }
 
 
-int generate_bezier_ts(struct bezier_t *in_bezier, struct t_t **out_t0, int *out_count, double *out_length)
+int generate_bezier_ts(struct bezier_t *in_bezier, struct t_t **out_t0, int *out_count, float *out_length)
 {
 	struct t_t *t0 = NULL;
 	ct0 = &t0;
@@ -212,7 +212,7 @@ void bezier_bigt_divide(struct t_t *t)
 }
 
 
-void generate_bezier_bigts(struct bezier_t *in_bezier, struct t_t **out_t0, int *out_count, double *out_length)
+void generate_bezier_bigts(struct bezier_t *in_bezier, struct t_t **out_t0, int *out_count, float *out_length)
 {
 	struct t_t *t0 = NULL;
 	ct0 = &t0;
